pyproject.toml
[tool.poetry]
name = "restframework"
version = "0.1.0"
description = ""
authors = ["Mrudhulraj <mrudhulraj.n@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
python-dotenv = "^0.20.0"
psycopg2-binary = "^2.9.3"
Django = "^4.0.4"
djangorestframework = "^3.13.1"
mysqlclient = "^2.1.0"
autopep8 = "^1.6.0"
pep8 = "^1.7.1"
pylint = "^2.13.9"
pylint-django = "^2.5.3"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
